refactor: split train helpers
