+++ bot_trade/tools/monitor_launch.py
@@ -8,6 +8,7 @@ import subprocess
 import platform
 import shlex
 from typing import List
+import logging
 from pathlib import Path
 
 ROOT = Path(__file__).resolve().parents[1]
@@ -107,3 +108,102 @@ def launch_new_console(title: str, module: str, args: List[str]) -> None:
     if debug:
         print("[MONITOR]", base_cmd)
     subprocess.Popen(base_cmd, cwd=ROOT)
+
+
+def spawn_monitor_manager(root_dir: str, args, run_id: str, headless: bool = True):
+    """Launch monitor_manager with minimal logic shared across callers."""
+    cmd = [
+        sys.executable,
+        "-m",
+        "bot_trade.tools.monitor_manager",
+        "--symbol",
+        args.symbol,
+        "--frame",
+        args.frame,
+        "--run-id",
+        run_id,
+        "--base",
+        root_dir,
+    ]
+    if headless:
+        cmd.append("--no-wait")
+
+    child_env = os.environ.copy()
+    child_env.setdefault("BOT_TRADE_ROOT", root_dir)
+
+    if not headless:
+        try:
+            proc = subprocess.Popen(cmd, env=child_env, shell=False)
+            logging.getLogger(__name__).info("[monitor] spawned (interactive)")
+            return proc
+        except Exception:  # pragma: no cover
+            logging.getLogger(__name__).warning("monitor launch failed")
+            return None
+
+    success = False
+    if os.name == "nt":
+        CREATE_NEW_CONSOLE = 0x00000010
+        DETACHED_PROCESS = 0x00000008
+        CREATE_NEW_PROCESS_GROUP = 0x00000200
+        CREATE_NO_WINDOW = 0x08000000
+
+        def _try_spawn_win(flags, close_fds):
+            return subprocess.Popen(
+                cmd,
+                stdin=subprocess.DEVNULL,
+                stdout=subprocess.DEVNULL,
+                stderr=subprocess.DEVNULL,
+                creationflags=flags,
+                close_fds=close_fds,
+                env=child_env,
+                shell=False,
+            )
+
+        try:
+            _try_spawn_win(CREATE_NO_WINDOW | CREATE_NEW_PROCESS_GROUP, True)
+            success = True
+        except Exception:
+            try:
+                _try_spawn_win(DETACHED_PROCESS | CREATE_NEW_PROCESS_GROUP, False)
+                success = True
+            except Exception:
+                try:
+                    pyw = sys.executable.replace("python.exe", "pythonw.exe")
+                    if os.path.exists(pyw):
+                        cmd2 = [pyw] + cmd[1:]
+                        subprocess.Popen(
+                            cmd2,
+                            stdin=subprocess.DEVNULL,
+                            stdout=subprocess.DEVNULL,
+                            stderr=subprocess.DEVNULL,
+                            creationflags=CREATE_NO_WINDOW | CREATE_NEW_PROCESS_GROUP,
+                            close_fds=True,
+                            env=child_env,
+                            shell=False,
+                        )
+                        success = True
+                    else:
+                        raise FileNotFoundError
+                except Exception:
+                    try:
+                        _try_spawn_win(CREATE_NEW_CONSOLE | CREATE_NEW_PROCESS_GROUP, True)
+                        success = True
+                    except Exception:
+                        success = False
+    else:
+        try:
+            subprocess.Popen(
+                cmd,
+                stdin=subprocess.DEVNULL,
+                stdout=subprocess.DEVNULL,
+                stderr=subprocess.DEVNULL,
+                env=child_env,
+                shell=False,
+                start_new_session=True,
+            )
+            success = True
+        except Exception:
+            success = False
+    if success:
+        logging.getLogger(__name__).info("[monitor] spawned (headless)")
+    return None
